Refactor accounts controller

+++ app/controllers/refinery/multisites/admin/accounts_controller.rb
@@ -13,9 +13,12 @@ module Refinery
         def create
           @account = Refinery::Multisites::Account.new(account_params)
           if @account.valid?
-            Apartment::Tenant.create(@account.subdomain)
-            @account.save
-            redirect_to refinery.multisites_admin_accounts_path
+            if Apartment::Tenant.create(@account.subdomain)
+              @account.save
+              redirect_to refinery.multisites_admin_accounts_path
+            else
+              render action: "new"
+            end
           else
             render action: "new"
           end
@@ -26,8 +29,9 @@ module Refinery
         end
 
         def destroy
-          current_account.destroy
-          Apartment::Tenant.drop(current_subdomain)
+          if Apartment::Tenant.drop(current_subdomain)
+            current_account.destroy
+          end
         end
 
         protected
